requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-dysmsapi==2.1.2
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.1.1
optionaldict==0.1.2
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wechatpy==1.8.18
xmltodict==0.14.2
